cleaned up rewrite content ui
using model picked by user..
using user selected temperature..

----

if the language selected is in user's lanaguage then the feedback is in user language at the skill level of the user.. if not the content will be in user selected language at the selected level but  the feedback will be in user base language.

if user base language is not available, then it will be defaulted to English at beginner level.